Several changes to initial condition, enhancement speed-up, the plotting.
Removed the initial condition applying function. I now do the initial
condition when I first initialize the solution object. I think this
makes more sense (and reduces the number of lines). I had to reformat
some of the tests to reflect this change.

I added a storage variable to the enhancement procedure to avoid the
vstack/append operations that result in building uf if you don't have
a temporary storage variable.

I modified the plotting script to make it a bit more generic/easier to
use in general.